Scan/requirements.txt
baidu_aip==2.2.18.0
numpy==1.21.5
opencv_python==4.5.5.62
